Demonstrate how to define execution environments
Added a Jenkinsfile that demonstrates how to define
an execution environment using Docker images and containers
to run inside a pipeline.

This allows the Pipeline to define the environment and tools
required without having to configure various system tools and
dependencies on agents manually and use practically any tool
which can be packaged in a Docker container.

For more agent specification options, consult the syntax reference.